move PostProcessorFrontendAccess to a separate file
